Resolving discussions from PR
