+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "python.formatting.provider": "black",
+  "python.pythonPath": "/home/michael/.virtualenvs/backend-9JyP9uwd/bin/python3.8"
+}
