+++ backend/.pre-commit-config.yaml
@@ -32,7 +32,7 @@ repos:
         exclude: test.*
     -   id: pytest
         name: pytest
-        entry: pytest
+        entry: pytest backend
         language: python
         always_run: true
         pass_filenames: false
