+++ backend/src/main.py
@@ -1,8 +1,10 @@
 from starlette.middleware.cors import CORSMiddleware
 from fastapi import FastAPI
-from routers.user_auth import login, register, get_token, utils
 
-utils.connect_to_db()
+from .routers.auth import auth_router
+from .auth.auth import connect_to_db
+
+connect_to_db()
 
 app = FastAPI()
 
@@ -19,6 +21,4 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-app.include_router(login.router)
-app.include_router(register.router)
-app.include_router(get_token.router)
+app.include_router(auth_router)
