+++ docker/docker-compose.yml
@@ -2,7 +2,7 @@ version: '3'
 services:
   backend:
     container_name: backend-container
-    build: 
+    build:
       context: ..
       dockerfile: docker/Dockerfile.backend
     command: pipenv run uvicorn src.main:app --host 0.0.0.0 --port 80 --reload
@@ -25,10 +25,10 @@ services:
       - '27017-27019:27017-27019'
   frontend:
     container_name: frontend-container
-    build: 
+    build:
       context: ..
       dockerfile: docker/Dockerfile.frontend
-    command: npm start  
+    command: npm start
     volumes:
       - '../frontend:/app'
       - '/app/node_modules'
