+++ frontend/public/index.html
@@ -31,7 +31,7 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    
+
     <div id="root"></div>
     <!--
       This HTML file is a template.
