+++ frontend/src/components/Board/Board.scss
@@ -11,7 +11,7 @@
     padding: 5px;
     margin-right: 4vw;
     height: 90%;
-}  
+}
 
 .sudoku table {
     border: 3px solid black;
