+++ frontend/src/components/Draggable/DragPanel/DragPanel.scss
@@ -4,7 +4,7 @@
     border-radius: 5px;
     background: $white;
     padding: 5px;
-}  
+}
 
 .dragPanel tbody{
     display: flex;
