+++ frontend/src/components/GameView/Timer/Timer.scss
